assets/server-template/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
strands-agents[bedrock]>=0.1.0
boto3>=1.35.0
pydantic>=2.0.0
